CurioNext_pvt_ltd/requirements.txt
pandas
matplotlib
requests
python-dotenv
groq
flask-cors
livekit-agents[groq,silero,turn-detector]~=1.0rc